More of general cleanup
- cosmetic style changes and refactoring
- get rid of nested directory
- remove dead code
- unexport matcher struct
- use canonical tag key
- update suite name
